move more rlearner arg cleanup into utils
explicitly list packaging library in requirements
