+++ causalml/inference/meta/rlearner.py
@@ -11,7 +11,7 @@ from sklearn.model_selection import cross_val_predict, KFold, train_test_split
 from xgboost import XGBRegressor
 
 from causalml.inference.meta.utils import (check_control_in_treatment, check_p_conditions,
-    clean_xgboost_objective)
+    get_xgboost_objective_metric)
 from causalml.inference.meta.explainer import Explainer
 
 logger = logging.getLogger('causalml')
@@ -545,22 +545,12 @@ class XGBRRegressor(BaseRRegressor):
                                                       (default = 'rank:pairwise')
             effect_learner_n_estimators (int, optional): number of trees to fit for the effect learner (default = 500)
         """
-        def clean_dict_keys(orig):
-            return {clean_xgboost_objective(k): v for (k, v) in orig.items()}
 
-        metric_mapping = clean_dict_keys({
-            'rank:pairwise': 'auc',
-            'reg:squarederror': 'rmse',
-        })
-
-        effect_learner_objective = clean_xgboost_objective(effect_learner_objective)
-
-        assert (effect_learner_objective in metric_mapping), \
-            'Effect learner objective must be one of: ' + ", ".join(metric_mapping)
         assert isinstance(random_state, int), 'random_state should be int.'
 
-        self.effect_learner_objective = effect_learner_objective
-        self.effect_learner_eval_metric = metric_mapping[effect_learner_objective]
+        objective, metric = get_xgboost_objective_metric(effect_learner_objective)
+        self.effect_learner_objective = objective
+        self.effect_learner_eval_metric = metric
         self.effect_learner_n_estimators = effect_learner_n_estimators
         self.early_stopping = early_stopping
         if self.early_stopping:
