+++ requirements.txt
@@ -17,3 +17,4 @@ shap
 dill
 lightgbm
 pygam
+packaging
